:ok_hand: Renaming Navigation Cmp
