+++ src/scenes/Receive/index.js
@@ -5,7 +5,7 @@ import { tint } from 'polished'
 import Feather from 'react-native-vector-icons/Feather'
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
 
-import NavigationHeader from '../../components/Navigation/NavigationHeader'
+import NavigationHeader from '../../components/Navigation/Header'
 import QRCode from '../../components/QRCode'
 import * as Utils from '../../components/Utils'
 import { Colors, FontSize } from '../../components/DesignSystem'
