+++ src/scenes/Seed/Create.js
@@ -4,7 +4,7 @@ import { ActivityIndicator, Alert } from 'react-native'
 import * as Utils from '../../components/Utils'
 import { Colors } from '../../components/DesignSystem'
 import ButtonGradient from '../../components/ButtonGradient'
-import NavigationHeader from '../../components/Navigation/NavigationHeader'
+import NavigationHeader from '../../components/Navigation/Header'
 
 import { getUserSecrets } from '../../utils/secretsUtils'
