+++ src/scenes/Settings/index.js
@@ -14,7 +14,7 @@ import { Colors, Spacing } from '../../components/DesignSystem'
 
 import { getUserSecrets } from '../../utils/secretsUtils'
 import fontelloConfig from '../../assets/icons/config.json'
-import NavigationHeader from '../../components/Navigation/NavigationHeader'
+import NavigationHeader from '../../components/Navigation/Header'
 
 const Icon = createIconSetFromFontello(fontelloConfig, 'tronwallet')
